requirements.txt
discord.py==2.6.1
Flask==2.3.3
